requirements.txt
Flask==2.3.3
supabase==1.1.1
python-dotenv==1.0.0